refactor(core/ui): move altcoin and webauthn layouts to separate file
Also fix types.
