+++ core/src/apps/webauthn/add_resident_credential.py
@@ -1,7 +1,9 @@
 import storage.device
 from trezor import wire
 from trezor.messages import Success, WebAuthnAddResidentCredential
-from trezor.ui.layouts import confirm_webauthn, show_error_and_raise
+from trezor.ui.components.common.webauthn import ConfirmInfo
+from trezor.ui.layouts import show_error_and_raise
+from trezor.ui.layouts.tt.webauthn import confirm_webauthn
 
 from .credential import Fido2Credential
 from .resident_credentials import store_resident_credential
