+++ core/src/apps/webauthn/remove_resident_credential.py
@@ -3,7 +3,7 @@ import storage.resident_credentials
 from trezor import wire
 from trezor.messages import Success, WebAuthnRemoveResidentCredential
 from trezor.ui.components.common.webauthn import ConfirmInfo
-from trezor.ui.layouts import confirm_webauthn
+from trezor.ui.layouts.tt.webauthn import confirm_webauthn
 
 from .credential import Fido2Credential
 from .resident_credentials import get_resident_credential
